Added basic access to CAN object dictionary with channel communication.
The current channel handling the CoE requests is c_coe_r (FIXME rename this),
the command objects are defined in coecmd.h which should be included into the
firmware app for easier access. Currently three commands are supported:
- CAN_GET_OBJECT <index|subindex>
	returns the current value of the object
- CAN_SET_OBJECT <index|subindex> <value>
	answers with the value written
- CAN_OBJECT_TYPE <index|subindex>
	returns type of the object

Note: <index|subindex> is one unsigned number where bits 23..8 define the index
and bits 7..0 set the subindex of the object.